Rename awslogs to cloudwatch
